tree: read TLS key in nvme_configure_ctrl()
nvme_configure_ctrl() is reading the values from sysfs, so we
should be reading the TLS key, too, if present.

+++ src/nvme/fabrics.h
@@ -57,7 +57,7 @@ struct nvme_fabrics_config {
 	int nr_poll_queues;
 	int tos;
 	int keyring;
-	int tls_key;
+	long tls_key;
 
 	bool duplicate_connect;
 	bool disable_sqflow;
